Refactor Groups helper objects
- Move data helpers and manager to separate modules
- Try to clean up exports to reduce redundancy between different
  `__init__.py` modules
- Ensure enums are exported at the top of the SDK, and name them all
  `Group*` so that their purpose/context is clear
- Add to doc page where missing
