+++ src/globus_sdk/__init__.py
@@ -19,20 +19,26 @@ from .exc import (
 )
 from .local_endpoint import LocalGlobusConnectPersonal
 from .response import GlobusHTTPResponse
-from .services import (
+from .services.auth import (
     AuthAPIError,
     AuthClient,
-    BatchMembershipActions,
     ConfidentialAppAuthClient,
-    DeleteData,
+    IdentityMap,
+    NativeAppAuthClient,
+)
+from .services.groups import (
+    BatchMembershipActions,
+    GroupMemberVisibility,
+    GroupRequiredSignupFields,
+    GroupRole,
     GroupsAPIError,
     GroupsClient,
     GroupsManager,
-    IdentityMap,
-    NativeAppAuthClient,
-    SearchAPIError,
-    SearchClient,
-    SearchQuery,
+    GroupVisibility,
+)
+from .services.search import SearchAPIError, SearchClient, SearchQuery
+from .services.transfer import (
+    DeleteData,
     TransferAPIError,
     TransferClient,
     TransferData,
@@ -70,6 +76,10 @@ __all__ = (
     "SearchQuery",
     "GroupsClient",
     "BatchMembershipActions",
+    "GroupMemberVisibility",
+    "GroupRequiredSignupFields",
+    "GroupRole",
+    "GroupVisibility",
     "GroupsManager",
     "LocalGlobusConnectPersonal",
 )
