--- src/globus_sdk/services/__init__.py
@@ -1,29 +1,0 @@
-from .auth import (
-    AuthAPIError,
-    AuthClient,
-    ConfidentialAppAuthClient,
-    IdentityMap,
-    NativeAppAuthClient,
-)
-from .groups import BatchMembershipActions, GroupsAPIError, GroupsClient, GroupsManager
-from .search import SearchAPIError, SearchClient, SearchQuery
-from .transfer import DeleteData, TransferAPIError, TransferClient, TransferData
-
-__all__ = (
-    "AuthClient",
-    "AuthAPIError",
-    "BatchMembershipActions",
-    "ConfidentialAppAuthClient",
-    "NativeAppAuthClient",
-    "IdentityMap",
-    "DeleteData",
-    "GroupsAPIError",
-    "GroupsClient",
-    "GroupsManager",
-    "SearchAPIError",
-    "SearchClient",
-    "SearchQuery",
-    "TransferAPIError",
-    "TransferClient",
-    "TransferData",
-)
