+++ src/globus_sdk/services/groups/__init__.py
@@ -1,9 +1,21 @@
-from .client import BatchMembershipActions, GroupsClient, GroupsManager
+from .client import GroupsClient
+from .data import (
+    BatchMembershipActions,
+    GroupMemberVisibility,
+    GroupRequiredSignupFields,
+    GroupRole,
+    GroupVisibility,
+)
 from .errors import GroupsAPIError
+from .manager import GroupsManager
 
 __all__ = (
     "GroupsClient",
     "GroupsAPIError",
     "GroupsManager",
     "BatchMembershipActions",
+    "GroupMemberVisibility",
+    "GroupRequiredSignupFields",
+    "GroupRole",
+    "GroupVisibility",
 )
